Add high-level specification for "state transfer".
In this specification, a mutator transfers state to a consumer via
snapshot and delta messages. The network ocassionally disconnects the
consumer from the mutator.

---

Properties:

1. The consumer should always eventually see the latest state the mutator
has picked, if we assume some fairness conditions.

2. The consumer should never accumulate an invalid state.

---

I've added a liveness property, i.e., a property that something good
should eventually happen, for (1) above. Actually, it is slightly more
strict! It says that the consumer eventually sees _any_ state the
mutator picks. TLC reports that this property is true for all of the
allowed behaviours of my specification; however, I don't yet understand
why this is the case. I've added a comment that describes a behaviour
prefix that I think invalidates the property. I suspect my understanding
of the "leads to" operator, `~>`, must be incorrect.