Show that SendDelta is broken.
I specify that

```tla
Mutations == [PickableValues -> PickableValues]
```

i.e., `Mutations` is the set of functions that (totally) relate one
pickable value to another.

However, `SendDelta` is only activated when there exists some `mutation`
in `Mutations` such that the domain of `mutation` is _only_ the
singleton set of the current `observer` state value.

```tla
\E mutation \in Mutations: DOMAIN mutation = { observer }
```

In this commit, I have added that "existence condition" to the `TypeOK`
invariant. It shows that it is always broken. In the next commit, I'll
change the definition for `Mutations`. I believe this will also
invalidate the temporal property `TransfersState`, which I was surprised
was working, given its strictness.