Fix TransferDelta so that it is "enableable".
Previously, it was never enabled, as I was asking for a element in the
set of `Mutations` that was not a member. In particular, one where the
DOMAIN was only the set including the mutator's current state.

Now, I've introduced `Invalid` to simulate a partial function and made
sure that `TransferDelta` can be enabled.

Fixing this action invalidated the temporal property `TransfersState`,
as I had previously expected. This invalidation allowed me to remove my
comment about being surprised that it was true for all allowed
behaviours.

I've weakened `TransfersState` property to expect either the observer to
eventually see the value the mutator picks or to see a disconnection,
i.e., `Empty`.